feat(Button): updated CSS dev pattern
